requirements.txt
fastapi
uvicorn
psycopg2
SQLAlchemy
python-dotenv
alembic
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pytest
requests
httpx
coverage


